Fix managed helpers and added missing ones (#77)
