+++ CMSIS-OS/ChibiOS/ST_NUCLEO144_F412ZG_NF/managed_helpers/README.md
@@ -2,4 +2,6 @@
 
 This folder contains helper classes to make developers life easier when coding for this target board.
 
-* [All](ST_NUCLEO144_F412ZG_NF.cs)
+* [Adc](ST_NUCLEO144_F412ZG_NF.Adc.cs)
+* [Gpio](ST_NUCLEO144_F412ZG_NF.Gpio.cs)
+* [LEDs](ST_NUCLEO144_F412ZG_NF.LEDs.cs)
