+++ CMSIS-OS/ChibiOS/ST_NUCLEO144_F412ZG_NF/managed_helpers/ST_NUCLEO144_F412ZG_NF.Adc.cs
@@ -1,42 +1,42 @@
 //
-// Copyright (c) 2018 The nanoFramework project contributors
+// Copyright (c) 2019 The nanoFramework project contributors
 // See LICENSE file in the project root for full license information.
 //
 
-namespace nanoFramework.Targets.ST_NUCLEO64_F401RE
+namespace nanoFramework.Targets.ST_NUCLEO144_F412ZG_NF
 {
     /// <summary>
-    /// Enumeration of ADC channels available on ST_NUCLEO64_F401RE
+    /// Enumeration of ADC channels available on board ST_NUCLEO144_F412ZG_NF
     /// </summary>
     public static class AdcChannels
     {
         /// <summary>
-        /// Channel 0, exposed as A0, connected to pin 1 on CN8 = PA0 (ADC1 - IN0)
+        /// Channel 0, exposed on CN9 Pin 1, connected to PA3 (ADC1 - IN3)
         /// </summary>
         public const int Channel_0 = 0;
 
         /// <summary>
-        /// Channel 1, exposed as A1, connected to pin 2 on CN8 = PA1 (ADC1 - IN1)
+        /// Channel 1, exposed on CN9 Pin 3, connected to PC0 (ADC1 - IN10)
         /// </summary>
         public const int Channel_1 = 1;
 
         /// <summary>
-        /// Channel 2, exposed as A2, connected to pin 3 on CN8 = PA4 (ADC1 - IN4)
+        /// Channel 2, exposed on CN9 Pin 5, connected to PC3 (ADC1 - IN13)
         /// </summary>
         public const int Channel_2 = 2;
 
         /// <summary>
-        /// Channel 3, exposed as A3, connected to pin 4 on CN8 = PB0 (ADC1 - IN8)
+        /// Channel 3, exposed on CN9 Pin 7, connected to PC1 (ADC1 - IN11)
         /// </summary>
         public const int Channel_3 = 3;
 
         /// <summary>
-        /// Channel 4, exposed as A4, connected to pin 5 on CN8 = PC1 (ADC1 - IN11)
+        /// Channel 4, exposed on CN9 Pin 9, connected to PC4 (ADC1 - IN14)
         /// </summary>
         public const int Channel_4 = 4;
 
         /// <summary>
-        /// Channel 5, exposed on A5, connected to pin 6 on CN8 = PC0 (ADC1 - IN10)
+        /// Channel 5, exposed on CN9 Pin 11, connected to PC5 (ADC1 - IN15)
         /// </summary>
         public const int Channel_5 = 5;
