+++ CMSIS-OS/ChibiOS/ST_NUCLEO144_F746ZG/managed_helpers/README.md
@@ -2,4 +2,4 @@
 
 This folder contains helper classes to make developers life easier when coding for this target board.
 
-* [Adc](Adc.cs)
+* [Adc](ST_NUCLEO144_F746ZG.Adc.cs)
