+++ CMSIS-OS/ChibiOS/ST_NUCLEO144_F746ZG/managed_helpers/ST_NUCLEO144_F746ZG.Adc.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2018 The nanoFramework project contributors
+// Copyright (c) 2019 The nanoFramework project contributors
 // See LICENSE file in the project root for full license information.
 //
