+++ CMSIS-OS/ChibiOS/ST_NUCLEO64_F401RE_NF/managed_helpers/README.md
@@ -2,4 +2,4 @@
 
 This folder contains helper classes to make developers life easier when coding for this target board.
 
-* [Adc](Adc.cs)
+* [Adc](ST_NUCLEO64_F401RE.Adc.cs)
