+++ CMSIS-OS/ChibiOS/ST_STM32F4_DISCOVERY/managed_helpers/README.md
@@ -2,4 +2,4 @@
 
 This folder contains helper classes to make developers life easier when coding for this target board.
 
-* [Adc](ST_NUCLEO64_F411RE_NF.Adc.cs)
+* [Adc](ST_STM32F4_DISCOVERY.Adc.cs)
